backend/requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
pydantic==1.8.1
python-dotenv==0.15.0
python-magic==0.4.22
python-multipart==0.0.5
PyYAML==5.4.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
